NoiseBase in lib-builtin-effects
